Add fetching tiles version. Clean up
